refactor: move collectible item components
