+++ src/components/Collectibles/CollectibleItemSummary.vue
